+++ plugins/pong.go
@@ -7,7 +7,7 @@ import (
 // Pong returns the string "Pong" when a user types "--Ping"
 func (r *Record) Pong(msg *dg.MessageCreate) string {
 	// Check the last time the user made this request
-	alertUser, tooSoon := r.checkLastAsk(msg)
+	alertUser, tooSoon := r.CheckLastAsk(msg)
 	if tooSoon {
 		return alertUser
 	}